+++ src/main/java/jsonvalues/UserError.java
@@ -26,142 +26,142 @@ public final class UserError extends RuntimeException {
 
     static UserError isNotAJsBool(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsBool expected, but %s was found",
+                                           String.format("`JsBool` expected, but `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isBool() before invoking asJsBool()"
+                                           "call the guard condition `isBool()` before invoking `asJsBool()`"
                                           ));
     }
 
     static UserError isNotAJsInt(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsInt expected, but %s was found",
+                                           String.format("`JsInt` expected, but `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isInt() before invoking asJsInt()"
+                                           "call the guard condition `isInt()` before invoking `asJsInt()`"
                                           ));
     }
 
     static UserError isNotAJsInstant(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsInstant expected or a JsString representing an instant formatted in ISO-8601, but %s was found",
+                                           String.format("`JsInstant` expected or a `JsString` representing an instant formatted in ISO-8601, but `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isInstant() before invoking asJsInstant()"
+                                           "call the guard condition `isInstant()` before invoking `asJsInstant()`"
                                           ));
     }
 
     static UserError isNotAJsBinary(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsBinary expected or a JsString representing an array of bytes encoded in base64, " +
+                                           String.format("`JsBinary` expected or a `JsString` representing an array of bytes encoded in base64, " +
                                                          "but %s was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isBinary() before invoking toJsBinary()"
+                                           "call the guard condition `isBinary()` before invoking `toJsBinary()`"
                                           ));
     }
 
     static UserError isNotAJsLong(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsLong expected, but %s was found",
+                                           String.format("`JsLong` expected, but `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isLong() before invoking toJsLong()"
+                                           "call the guard condition `isLong()` before invoking `toJsLong()`"
                                           ));
     }
 
     static UserError isNotAJsDouble(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsDouble expected, but %s was found",
+                                           String.format("`JsDouble` expected, but` `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isDouble() before invoking toJsDouble()"
+                                           "call the guard condition `isDouble()` before invoking `toJsDouble()`"
                                           ));
     }
 
     static UserError isNotAJsBigInt(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsBigInt expected, but %s was found",
+                                           String.format("`JsBigInt` expected, but `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isBigInt() or isIntegral() before invoking toJsBigInt()"
+                                           "call the guard condition `isBigInt()` or `isIntegral()` before invoking `toJsBigInt()`"
                                           ));
     }
 
     static UserError isNotAJsBigDec(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsBigDec expected, but %s was found",
+                                           String.format("`JsBigDec` expected, but `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isBigDec() or isDecimal() before invoking toJsBigDec()"
+                                           "call the guard condition `isBigDec()` or `isDecimal()` before invoking `toJsBigDec()`"
                                           ));
     }
 
     static UserError isNotAJsArray(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsArray expected, but %s was found",
+                                           String.format("`JsArray` expected, but `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isArray() before invoking toJsArray()"
+                                           "call the guard condition `isArray()` before invoking `toJsArray()`"
                                           ));
     }
 
     static UserError isNotAJsPrimitive(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsPrimitive expected, but %s was found",
+                                           String.format("`JsPrimitive` expected, but `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isPrimitive() before invoking toJsPrimitive()"
+                                           "call the guard condition `isPrimitive()` before invoking `toJsPrimitive()`"
                                           ));
     }
 
     static UserError isNotAJsNumber(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsNumber expected, but %s was found",
+                                           String.format("`JsNumber` expected, but `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isNumber() before invoking toJsArray()"
+                                           "call the guard condition `isNumber()` before invoking `toJsArray()`"
                                           ));
     }
 
     static UserError isNotAJsObj(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsObj expected, but %s was found",
+                                           String.format("`JsObj` expected, but `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isObj() before invoking toJsObj()"
+                                           "call the guard condition `isObj()` before invoking `toJsObj()`"
                                           ));
     }
 
     static UserError isNotAJson(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("Json expected, but %s was found",
+                                           String.format("`Json` expected, but `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isJson() or isArray() or isObj() before invoking toJson()"
+                                           "call the guard condition `isJson()` or `isArray()` or `isObj()` before invoking toJson()"
                                           ));
     }
 
     static UserError isNotAJsString(final JsValue elem) {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           String.format("JsStr expected, but %s was found",
+                                           String.format("`JsStr` expected, but `%s` was found",
                                                          elem.getClass()
                                                         ),
-                                           "call the guard condition isStr() before invoking toJsStr()"
+                                           "call the guard condition `isStr()` before invoking`toJsStr()`"
                                           ));
     }
 
     static UserError lastOfEmptyPath() {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           "last() of empty path",
-                                           "call the guard condition isEmpty() before invoking last()"
+                                           "`last()` of empty path",
+                                           "call the guard condition `isEmpty()` before invoking `last()`"
                                           ));
     }
 
     static UserError headOfEmptyPath() {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           "head() of empty path",
-                                           "call the guard condition isEmpty() before invoking head()"
+                                           "`head()` of empty path",
+                                           "call the guard condition `isEmpty()` before invoking `head()`"
                                           ));
     }
 
@@ -177,44 +177,44 @@ public final class UserError extends RuntimeException {
 
     static UserError tailOfEmptyPath() {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           "tail() of empty path",
-                                           "call the guard condition isEmpty() before invoking tail()"
+                                           "`tail()` of empty path",
+                                           "call the guard condition `isEmpty()` before invoking `tail()`"
                                           ));
     }
 
     static UserError initOfEmptyPath() {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           "init() of empty path",
-                                           "call the guard condition isEmpty() before invoking init()"
+                                           "`init()` of empty path",
+                                           "call the guard condition `isEmpty()` before invoking `init()`"
                                           ));
     }
 
 
     static UserError asKeyOfIndex() {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           "asKey() of index",
-                                           "use the guard condition position.isKey() before"
+                                           "`asKey()` of index",
+                                           "use the guard condition `position.isKey()` before"
                                           ));
     }
 
     static UserError asIndexOfKey() {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           "asIndex() of key",
-                                           "use the guard condition position.isIndex() before"
+                                           "`asIndex()` of key",
+                                           "use the guard condition `position.isIndex()` before"
                                           ));
     }
 
     static UserError incOfKey() {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           "inc() of key",
-                                           "use the guard condition last().isIndex() before invoking inc()"
+                                           "`inc()` of key",
+                                           "use the guard condition `last().isIndex()` before invoking `inc()`"
                                           ));
     }
 
     static UserError decOfKey() {
         return new UserError(String.format(GENERAL_MESSAGE,
-                                           "dec() of key",
-                                           "use the guard condition last().isIndex() before invoking dec()"
+                                           "`dec()` of key",
+                                           "use the guard condition `last().isIndex()` before invoking `dec()`"
                                           ));
     }