+++ src/test/java/jsonvalues/api/TestJsObj.java
@@ -1139,4 +1139,5 @@ public class TestJsObj {
                                    .map(it -> it.key())
                                    .reduce("", (a, b) -> a + b));
     }
+
 }